+++ src/models/options.ts
@@ -1,18 +1,21 @@
-import {HomebridgeNeatoPlatform} from "../homebridgeNeatoPlatform";
-import {PlatformAccessory, PlatformConfig} from "homebridge";
-
 export class Options
 {
 	public eco: boolean;
 	public extraCare: boolean;
 	public noGoLines: boolean;
-	public spot: any;
+	public spotCharacteristics: boolean;
+	public spotRepeat: boolean;
+	public spotWidth: number;
+	public spotHeight: number;
 	
 	constructor()
 	{
 		this.eco = false;
 		this.extraCare = false;
 		this.noGoLines = false;
-		this.spot = {};
+		this.spotCharacteristics = false;
+		this.spotRepeat = false;
+		this.spotWidth = 200;
+		this.spotHeight = 200;
 	}
 }